Beta Version

+++ Spotify.NetStandard/Spotify.NetStandard/Requests/TuneableTrack.cs
@@ -23,7 +23,7 @@ namespace Spotify.NetStandard.Requests
         /// The duration of the track in milliseconds.
         /// </summary>
         [Description("duration_ms")]
-        public int? Duration { get; set; }
+        public long? Duration { get; set; }
 
         /// <summary>
         /// Energy is a measure from 0.0 to 1.0 and represents a perceptual measure of intensity and activity
